ai-playground/OCR-Doc-Analysis/requirements.txt
openai>=1.12.0
pydantic>=2.5.0
opencv-python>=4.9.0
Pillow>=10.0.0
numpy>=1.26.0

